scripts: integrate build script
